ecommerce/backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
psycopg2==2.8.6